move pd client to pd package
